SF patch #682514, mmapmodule.c write fix for LP64 executables
Make length an int so we get the right value from
PyArg_ParseTuple(args, "s#", &str, &length)

+++ Modules/mmapmodule.c
@@ -292,7 +292,7 @@ static PyObject *
 mmap_write_method(mmap_object *self,
 		  PyObject *args)
 {
-	long length;
+	int length;
 	char *data;
 
 	CHECK_VALID(NULL);
